gaia/validator/database/requirements.txt
# Database dependencies
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
aioredis>=2.0.0
alembic>=1.13.0  # For database migrations
psycopg2-binary>=2.9.9  # For PostgreSQL support
redis>=5.0.0  # For Redis support

# Prefect integration
prefect>=2.14.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0 